requirements.txt
matplotlib
google-api-core==2.11.1
google-auth==2.21.0
google-cloud-bigquery==3.11.3
google-cloud-core==2.3.2
google-cloud-resource-manager==1.10.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
google-cloud-discoveryengine
PyYAML==6.0.1
langsmith==0.0.7
langchain==0.0.235
google-cloud-core==2.3.2
SQLAlchemy==1.4.49
sqlalchemy-bigquery==1.7.0
google-cloud-bigquery-storage==2.22.0
sqlparse==0.4.4
pandas==2.0.3
google-cloud-aiplatform >= 1.31.0
google-cloud-logging
streamlit
streamlit-chat